Adding middleware #18
Saw the concept of middleware functions. What are they? and why use them?
Created two middleware function. The first one write a text to a console and passes the request to the next middleware
The second middleware write a text to the console and sends a response to the client (web browser)

Closes #18