Adding  layout to handlebars
changed the main.handlebars file to  apply css  styles conditionally
moved the header markup from shop and add-product page to the main.handlebars file. By doing this we can change the header once and it will affect all the pages that use this layout.

removed the header section from the add-product and shop

inside the shop.js and admin.js route handler export some properties to be use in conditions statement in the layout file