Working with Layout and Partials
created head.ejs that contains chunck of code that represent the head section of every view file.

created navigation.ejs that contains the chunck of code that represent the navigation section of every view file
created end.ejs that contains the chunck of code that represent the end section of every view file

change the 404.ejs file by removing the head section and replacing it by the ejs include statement
repeated the same thing  for add-product, shop view files.