Adding Product Model
Integrate the M "Model" part of the MVC pattern we are implementing.

created the Product model
used the Product model in the product controller file to save a Product in the postAddProduct controller fonction and to get product

Storing model data in files and fetching from file via the model.
refactoring file storage

closes #65