Storing Product Data and Displaying Product Data
Modified the add-product form so that we can also store the product image, the product price, and product description.
Modified the product model to incorporate, in addition to the title, the imageURL, the price and the product description
Modified the postAddProduct controller to get the product data from the add-product form and to save them
Add content to the views/shop/index.ejs to display the product stored in the json file.

closes #74