Adding the product ID to the path ---
We want to be able to display a product detail from the product list page.
For that we added a link with pointing to a /product/productid
For that link to work, we created a route handler for GET /products/:productId where :productId is the placeholder for the product ID.
In the product model, we refactor the class Product so that it created a id product property and saves a random number to it.

closes #81